chore: moving legacy redirect to utils directory
